+++ src/lib.rs
@@ -1,5 +1,14 @@
 //! Decoding and Encoding of WebP Images
 //!
+//! Copyright (C) 2025 Imazen LLC
+//!
+//! This program is free software: you can redistribute it and/or modify
+//! it under the terms of the GNU Affero General Public License as published
+//! by the Free Software Foundation, either version 3 of the License, or
+//! (at your option) any later version.
+//!
+//! For commercial licensing inquiries: support@imazen.io
+//!
 //! This crate provides both encoding and decoding of WebP images.
 //!
 //! # Features